+++ datagrowth/management/commands/grow_dataset.py
@@ -66,14 +66,3 @@ class Command(DatasetCommand):
             if timer >= timeout:
                 message = "Grow dataset command exceeded timeout"
                 log.error(message)
-
-    ###################################
-    # Legacy Community compatability
-    ###################################
-
-    def handle_community(self, dataset, *args, **options):
-        dataset.config = {"asynchronous": False}  # TODO: this is weird syntax as it is actually performing an update
-        dataset.save()
-        dataset.grow(*args)
-        log.info("Result: {}".format(dataset.kernel))
-        log.info("Growth: {}".format([growth.id for growth in dataset.growth_set.all()]))