+++ datagrowth/management/commands/load_dataset.py
@@ -1,14 +1,9 @@
 import logging
 import os
-import json
 import re
 
-from django.apps import apps
-from django.core.management.base import CommandError
-from django.contrib.contenttypes.models import ContentType
-
 from datagrowth.management.base import DatasetCommand
-from datagrowth.utils import get_dumps_path, objects_from_disk
+from datagrowth.utils import get_dumps_path
 
 
 log = logging.getLogger("datagrowth.command")
@@ -22,7 +17,6 @@ class Command(DatasetCommand):
     def add_arguments(self, parser):
         super().add_arguments(parser)
         parser.add_argument('-r', '--replace', action="store_true")
-        parser.add_argument('-t', '--transform-community', action="store_true")
 
     def get_dataset(self):  # picks the correct dataset from all available datasets based on signature
         for dataset in self.get_datasets():
@@ -45,57 +39,3 @@ class Command(DatasetCommand):
     def handle_dataset(self, dataset, *args, **options):
         replace = options["replace"]
         dataset.from_file(dataset.file_path, replace=replace)
-
-    ###################################
-    # Legacy Community compatability
-    ###################################
-
-    def bulk_create_community_objects(self, objects, transform_community):
-
-        obj = objects[0]
-        model = type(obj)
-
-        if transform_community:
-            if isinstance(obj, self.Individual):
-                model = self.Document
-                for obj in objects:
-                    collection_id = obj.collective_id if obj.collective_id else None
-                    obj.__class__ = model
-                    obj.collection_id = collection_id
-                    if isinstance(obj.properties, str):
-                        obj.properties = json.loads(obj.properties)
-            elif isinstance(obj, self.Collective):
-                model = self.Collection
-                for obj in objects:
-                    obj.__class__ = model
-            elif isinstance(obj, self.Growth):
-                for obj in objects:
-                    obj.input_type = ContentType.objects.get_for_model(
-                        self.Document if isinstance(obj.input, self.Individual) else self.Collection
-                    )
-                    obj.output_type = ContentType.objects.get_for_model(
-                        self.Document if isinstance(obj.output, self.Individual) else self.Collection
-                    )
-            else:
-                obj.kernel_type = ContentType.objects.get_for_model(
-                    self.Document if isinstance(obj.kernel, self.Individual) else self.Collection
-                )
-
-        model.objects.bulk_create(objects)
-
-    def handle_community(self, community, *args, **options):
-        transform_community = options.get("transform_community", False)
-        if transform_community:
-            log.info("Using transformation to change all data storage into Document and Collection")
-            self.Growth = apps.get_model("core", "Growth")
-            self.Document = apps.get_model(community._meta.app_label, "Document")
-            self.Individual = apps.get_model("core", "Individual")
-            self.Collection = apps.get_model(community._meta.app_label, "Collection")
-            self.Collective = apps.get_model("core", "Collective")
-        if not os.path.exists(community.file_path):
-            message = f"Dump with signature {community.signature} does not exist"
-            log.error(message)
-            raise CommandError(message)
-        with open(community.file_path, "r") as dump_file:
-            for objects in objects_from_disk(dump_file):
-                self.bulk_create_community_objects(objects, transform_community)