+++ datagrowth/processors/__init__.py
@@ -1,4 +1,4 @@
-from datagrowth.processors.base import Processor, QuerySetProcessor, ProcessorFactory
+from datagrowth.processors.base import Processor, ProcessorFactory
 from datagrowth.processors.growth import GrowthProcessor
 from datagrowth.processors.input import (ExtractProcessor, TransformProcessor, HttpSeedingProcessor,
                                          SeedingProcessorFactory)